fix: address code review findings — total_hours type, failed status, event loop deprecation
- Change SimulationConfig.total_hours from int to float (supports fractional hours)
- Set simulation status to 'failed' on runner exception instead of leaving 'running'
- Replace deprecated asyncio.get_event_loop() with asyncio.get_running_loop()

+++ src/forkcast/simulation/models.py
@@ -40,7 +40,7 @@ class AgentProfile:
 class SimulationConfig:
     """Generated simulation parameters."""
 
-    total_hours: int
+    total_hours: float
     minutes_per_round: int
     peak_hours: list[int]
     off_peak_hours: list[int]
